+++ examples/create-batch-proposal/index.js
@@ -1,6 +1,6 @@
 require('dotenv').config();
 
-const { Platform } = require('@openzeppelin/platform-sdk');
+const { Platform } = require('@openzeppelin/defender-sdk');
 
 const ERC20Token = '0x24B5C627cF54582F93eDbcF6186989227400Ac75';
 const RolesContract = '0xa50d145697530e8fef3F59a9643c6E9992d0f30D';