+++ examples/create-monitor/index.js
@@ -2,7 +2,7 @@
 require('dotenv').config();
 
 const abi = require('./abis/erc721.json');
-const { Platform } = require('@openzeppelin/platform-sdk');
+const { Platform } = require('@openzeppelin/defender-sdk');
 
 async function main() {
   const creds = { apiKey: process.env.API_KEY, apiSecret: process.env.API_SECRET };