+++ examples/upgrade-contract/CHANGELOG.md
@@ -1,11 +1,11 @@
-# @openzeppelin/platform-sdk-example-upgrade-contract
+# @openzeppelin/defender-sdk-example-upgrade-contract
 
 ## 0.2.1
 
 ### Patch Changes
 
 - Updated dependencies [efefd68]
-  - @openzeppelin/platform-sdk@0.3.0
+  - @openzeppelin/defender-sdk@0.3.0
 
 ## 0.2.0
 
@@ -24,4 +24,4 @@
 ### Patch Changes
 
 - Updated dependencies [496f88d]
-  - @openzeppelin/platform-sdk@0.2.0
+  - @openzeppelin/defender-sdk@0.2.0